JWT authentication Added

+++ controllers/notes.js
@@ -7,14 +7,13 @@ exports.create = async (req, res) => {
 		// Create a Note
 		const note = new Note({
 			title: req.body.title || "Untitled Note",
+			userId: req.user._id,
 			...req.body,
 		});
 		// Save Note in the database
 		const notes = await note.save();
 		return res.status(200).send({
 			message: "Note is created successfully",
-			status: 200,
-			code: "OK",
 			notes,
 		});
 	} catch (error) {
@@ -28,7 +27,9 @@ exports.create = async (req, res) => {
 // Retrieve and return all notes from the database.
 exports.findAll = async (req, res) => {
 	try {
-		const allNotes = await Note.find({});
+		const allNotes = await Note.find({
+			userId: req.user._id
+		});
 
 		return res.status(200).send({
 			message: "Notes are found successfully",
@@ -51,7 +52,10 @@ exports.findOne = async (req, res) => {
 				message: "Invalid Request Params"
 			})
 		}
-		const note = await Note.findById(req.params.noteId);
+		const note = await Note.findOne({
+			_id: req.params.noteId,
+			userId: req.user._id,
+		});
 		if (!note) {
 			return res.status(404).send({
 				message: "Given note is not found, please try again"
@@ -86,7 +90,15 @@ exports.update = async (req, res) => {
 				message: "Note content can not be empty"
 			});
 		}
-
+		const note = await Note.findOne({
+			_id: req.params.noteId,
+			userId: req.user._id
+		});
+		if (!note) {
+			return res.status(404).send({
+				message: "Given note is not found, please try again"
+			})
+		}
 		// Find note and update it with the request body
 		const updatedNote = await Note.findByIdAndUpdate(req.params.noteId, {
 			title: req.body.title || "Untitled Note",
@@ -116,6 +128,15 @@ exports.delete = async (req, res) => {
 		if (!req.params.noteId) {
 			return res.status(500).send({
 				message: "Invalid Request Params"
+			})
+		}
+		const note = await Note.findOne({
+			_id: req.params.noteId,
+			userId: req.user._id
+		});
+		if (!note) {
+			return res.status(404).send({
+				message: "Given note is not found, please try again"
 			})
 		}
 		const deletedNote = await Note.findByIdAndRemove(req.params.noteId);
